Separate schoool service in a directory
